chore: Moving ResponsiveLayout to helpers
