+++ src/Uno.Toolkit.UI/Controls/ResponsiveView/ResponsiveView.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
 using Windows.Foundation;
-
+using Uno.Toolkit.UI.Helpers;
 #if IS_WINUI
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
